+++ src/lib/core/interpolate.ts
@@ -1,8 +1,10 @@
 import { type KargsType } from "../types/tag.types";
 import {
+    type IZodTagRenderable,
     type IRenderable,
+    type IRenderableResult,
 } from "./renderable"
-import { collectChunks } from "./chunks";
+import { collectChunks, } from "./chunks";
 import { interpolateChunks } from "./interpolate-chunks";
 
 /**
@@ -10,6 +12,8 @@ import { interpolateChunks } from "./interpolate-chunks";
  * @param renderable target renderable
  * @param karg Keyword arguments object
  */
-export function interpolate<K extends KargsType>(renderable: IRenderable<K, any>, karg: K) {
-    return collectChunks(interpolateChunks(renderable, karg))
+export function interpolate<R extends IRenderable<K, O>, K extends KargsType, O extends any[]>(renderable: R, karg: K) {
+    const r = renderable as any as IZodTagRenderable
+    if (r.__static) return r.__static as IRenderableResult<R>
+    return collectChunks(interpolateChunks(r, karg)) as IRenderableResult<R>
 }